Reflected and decided that actually further refinement of structure made sense. The update events has been moved to event controllers. Instead of deno serve, created a post route that can be pinged via cron job with the correct header. Created an api example also. The generate, save and update / cron job functions can be tested via this.
